Refactored the control panel and workspace reload
I Used BroadcastChannel to trigger an event through browser tabs and/or windows if a challenge (re)started.
VSCode reloads the page.
noVNC changes the iframe src.

I added a description to the control panel and moved the notifications to the bottom of the control panel. I Limited the max-size to 50vh and made it scrollable. Therefor I changed the scrollbar to be more visible.

I added the description to the test of active_module.