Fix Accordion Item Overexpansion
When expanding an accordion item, the height of the SSH hint is being taken into consideration.
However, the SSH hint is not displayed when the item finishes expanding, so the item overshoots
before snapping to the correct height. By setting the hint to not display when inside a hidden
challenge workspace, the item does not incorrectly add the height of the hint.

So, whoops, this one was my fault (introduced in 0bf38c5).